# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: LINKa Identity API
  version: 0.1.0
  description: >-
    Internal API for optional accounts, anonymous installations, encrypted email
    identities, organizations, consent, telemetry preference, privacy requests,
    and product-scoped tokens. Workload credentials are role- and product-scoped.
    Raw email and root UUIDs never appear in a response or token.
servers:
  - url: http://localhost:8080
security:
  - internalBearer: []
paths:
...
  /v1/installations:
    post:
      operationId: createInstallation
      summary: Create an anonymous product installation and return only its pairwise key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallationRequest"
      responses:
        "201":
          description: Installation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/email-identities:
    post:
      operationId: registerEmailIdentity
      summary: Consume a verified email challenge and create or resolve an identity
      description: >-
        Product scope is the default. Donation namespace is always product-local.
        Global minor linkage fails unless the deployment feature flag is enabled;
        unknown age can never request global linkage.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailIdentityRequest"
      responses:
        "200":
          description: Existing identity resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailIdentityResponse"
        "201":
          description: Identity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailIdentityResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/email-verifications:
    post:
      operationId: beginEmailVerification
      summary: Encrypt email and create a short-lived pending ownership challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailVerificationRequest"
      responses:
        "202":
          description: Verification challenge created; no email is returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailVerificationResponse"
  /v1/internal/email-verifications/{id}/verify:
    post:
      operationId: verifyEmailOwnership
      summary: Record external ownership evidence as an email_verifier workload
      parameters:
        - $ref: "#/components/parameters/UUIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [product_id, evidence_id]
              properties:
                product_id:
                  $ref: "#/components/schemas/ProductID"
                evidence_id:
                  type: string
                  minLength: 1
                  maxLength: 200
      responses:
        "200":
          description: Ownership evidence recorded
  /v1/tokens:
    post:
      operationId: issueToken
      summary: Issue a short-lived product-scoped JWT with no email or root UUID
      description: Account JWTs contain the account subject key and a separate pairwise person_key for privacy aggregation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "201":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/organization-submissions:
    post:
      operationId: createOrganizationSubmission
      summary: Submit a free-text organization name for review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationSubmissionRequest"
      responses:
        "201":
          description: Submission created with pending status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/internal/organization-submissions/{id}/resolve:
    post:
      operationId: resolveOrganizationSubmission
      summary: Match or reject a pending organization submission
      parameters:
        - $ref: "#/components/parameters/UUIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationResolutionRequest"
      responses:
        "200":
          description: Submission resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/memberships:
    post:
      operationId: createMembership
      summary: Create or update a person's canonical organization membership
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipRequest"
      responses:
        "201":
          description: Membership recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    internalBearer:
      type: http
      scheme: bearer
      description: Per-workload credential. Server-side roles and product scopes authorize each route.
  parameters:
    UUIDPath:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
...
  responses:
    BadRequest:
      description: Request failed strict validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Internal bearer credential is missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Linkage policy forbids the operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Existing resource conflicts with the supplied data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    UUID:
      type: string
      format: uuid
    ProductID:
      type: string
      pattern: "^[a-z][a-z0-9_-]{1,63}$"
    OpaqueKey:
      type: string
      pattern: "^[a-f0-9]{64}$"
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CreatedStatus:
      type: object
      required: [id, status]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        status:
          type: string
    SubjectFields:
      type: object
      required: [subject_type, subject_key]
      properties:
        subject_type:
          type: string
          enum: [person, account, installation]
        subject_key:
          $ref: "#/components/schemas/OpaqueKey"
    InstallationRequest:
      type: object
      additionalProperties: false
      required: [platform]
      properties:
        product_id:
          $ref: "#/components/schemas/ProductID"
        platform:
          type: string
          minLength: 1
          maxLength: 64
    Installation:
      type: object
      required: [installation_key, product, platform, created_at]
      properties:
        installation_key:
          $ref: "#/components/schemas/OpaqueKey"
        product:
          $ref: "#/components/schemas/ProductID"
        platform:
          type: string
        created_at:
          type: string
          format: date-time
    EmailVerificationRequest:
      type: object
      additionalProperties: false
      required: [email, identity_namespace, age_category]
      properties:
        product_id:
          $ref: "#/components/schemas/ProductID"
        email:
          type: string
          format: email
          maxLength: 320
          writeOnly: true
        identity_namespace:
          type: string
          enum: [account, donation]
        age_category:
          type: string
          enum: [unknown, adult, minor]
        guardian_relationship:
          type: string
          minLength: 1
          maxLength: 120
          description: Optional free text for a minor; it is not proof of authority or consent.
        link_across_products:
          type: boolean
          default: false
        installation_key:
          $ref: "#/components/schemas/OpaqueKey"
    EmailVerificationResponse:
      type: object
      additionalProperties: false
      required: [verification_id, status, expires_at]
      properties:
        verification_id:
          $ref: "#/components/schemas/UUID"
        status:
          const: pending
        expires_at:
          type: string
          format: date-time
    EmailIdentityRequest:
      type: object
      additionalProperties: false
      required: [verification_id]
      properties:
        product_id:
          $ref: "#/components/schemas/ProductID"
        verification_id:
          $ref: "#/components/schemas/UUID"
        create_account:
          type: boolean
          default: false
    EmailIdentityResponse:
      type: object
      additionalProperties: false
      required: [subject_type, subject_key]
      properties:
        subject_type:
          type: string
          enum: [person, account]
        subject_key:
          $ref: "#/components/schemas/OpaqueKey"
    TokenRequest:
      type: object
      additionalProperties: false
      required: [subject_type, subject_key]
      properties:
        product_id:
          $ref: "#/components/schemas/ProductID"
        subject_type:
          type: string
          enum: [account, installation]
        subject_key:
          $ref: "#/components/schemas/OpaqueKey"
        ttl_seconds:
          type: integer
          minimum: 0
          description: Zero or omission uses configured short default; configured maximum always applies.
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_at]
      properties:
        access_token:
          type: string
        token_type:
          const: Bearer
        expires_at:
          type: string
          format: date-time
    OrganizationSubmissionRequest:
      allOf:
        - $ref: "#/components/schemas/SubjectFields"
        - type: object
          required: [product_id, submitted_name]
          properties:
            product_id:
              $ref: "#/components/schemas/ProductID"
            submitted_name:
              type: string
              minLength: 1
              maxLength: 240
      unevaluatedProperties: false
    OrganizationResolutionRequest:
      type: object
      additionalProperties: false
      required: [status, audit_note]
      properties:
        status:
          type: string
          enum: [matched, rejected]
        organization_id:
          $ref: "#/components/schemas/UUID"
        audit_note:
          type: string
          minLength: 1
          maxLength: 1000
...
    MembershipRequest:
      type: object
      additionalProperties: false
      required: [person_key, organization_id, product_id, status]
      properties:
        person_key:
          $ref: "#/components/schemas/OpaqueKey"
        organization_id:
          $ref: "#/components/schemas/UUID"
        product_id:
          $ref: "#/components/schemas/ProductID"
        role_label:
          type: string
          minLength: 1
          maxLength: 120
        status:
          type: string
          enum: [active, inactive, pending]
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
...